+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Remove 'no data' message displayed when saving the item. Fixes STSMACOM-384.
 * Fix Custom Fields multiselect Label field not read as required. Fixes STSMACOM-363.
 * Make Notes list column headings sortable. Refs STSMACOM-328.
+* Refactor from `bigtest/mirage` to `miragejs`.
 
 ## [4.2.0] (IN PROGRESS)